+++ src/api/friend.js
@@ -1,18 +1,10 @@
-import axios from 'axios';
-import { BASE_URL } from '../shared/constants';
-import { getAccessToken } from './getToken';
+import instance from './axios';
 
 const RESOURCE = '/users/friend';
 
 export const getFriendList = async () => {
   try {
-    const accesstoken = await getAccessToken();
-    const response = await axios.get(`${BASE_URL}${RESOURCE}`, {
-      headers: {
-        Authorization: accesstoken,
-      },
-    });
-
+    const response = await instance.get(RESOURCE);
     return response.data;
   } catch (error) {
     console.error('getFriendList Axios Error', error);
